Setup more services, repositories, utilities and aggregates

+++ src/application/services/like/LikeService.java
@@ -1,7 +1,8 @@
 package src.application.services.like;
 
+import src.application.providers.SessionProvider;
 import src.domain.entities.Like;
-import src.domain.entities.Post;
+import src.domain.aggregates.Post;
 import src.domain.entities.User;
 import src.infrastructure.repositories.like.LikeRepository;
 
@@ -14,23 +15,28 @@ public class LikeService {
      * The repository for handling like data.
      */
     private final LikeRepository likeRepository;
+    private final SessionProvider sessionProvider;
 
     /**
      * Constructs a LikeService instance, initializing the LikeRepository.
      */
-    public LikeService() {
+    public LikeService(SessionProvider sessionProvider) {
         this.likeRepository = new LikeRepository();
+        this.sessionProvider = sessionProvider;
     }
 
     /**
      * Likes a post on behalf of a user.
      *
-     * @param user The user who is liking the post.
      * @param post The post to be liked.
      */
-    public void likePost(User user, Post post) {
-        Like like = new Like(user, post);
-        likeRepository.save(like);
+    public void likePost(Post post) {
+        if (sessionProvider.isAuthenticated()) {
+            Like like = new Like(sessionProvider.getAuthenticatedUser(), post);
+            likeRepository.save(like);
+        } else {
+            // Handle not authenticated
+        }
     }
 
     /**
@@ -39,6 +45,16 @@ public class LikeService {
      * @param like The like entity representing the like to be removed.
      */
     public void unlikePost(Like like) {
-        likeRepository.delete(like.getId());
+        if (sessionProvider.isAuthenticated()) {
+            User likeUser = like.getUser();
+            User authenticatedUser = sessionProvider.getAuthenticatedUser();
+            if (!likeUser.getId().equals(authenticatedUser.getId())) {
+                throw new IllegalArgumentException("User does not have permission to remove this like.");
+            } else {
+                likeRepository.delete(like.getId());
+            }
+        } else {
+            // Handle not authenticated
+        }
     }
 }
